Move order maker/taker balance/allowance setting to a scenario module and use in both orderValidation and integration tests
